pyproject.toml
[tool.poetry]
name = "studere"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.115.3"
sqlalchemy = "2.0.36"
starlette = "0.41.0"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
exceptiongroup = "1.2.2"
greenlet = "3.1.1"
idna = "3.10"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
pyjwt = "2.9.0"
sniffio = "1.3.1"
typing-extensions = "4.12.2"
python-decouple = "^3.8"
uvicorn = "^0.32.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
